requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
argilla==0.0.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
cohere==4.11.2
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.9
dnspython==2.3.0
duckdb==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.99.1
ffmpy==0.3.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.35.2
gradio_client==0.2.7
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.223
langchainplus-sdk==0.0.20
linkify-it-py==2.0.2
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.0
olefile==0.46
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.1
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
Pillow==10.0.0
pinecone-client==2.2.2
portalocker==2.7.0
posthog==3.0.1
protobuf==4.23.3
pulsar-client==3.2.0
pycparser==2.21
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
PyMuPDF==1.22.5
pypandoc==1.11
pyparsing==3.1.0
pypdf==3.12.0
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
qdrant-client==1.3.1
redis==4.6.0
regex==2023.6.3
requests==2.31.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
SQLAlchemy==2.0.17
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
unstructured==0.7.12
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
